requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file requirements.txt
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
anyio=3.5.0=py37h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h7f8727e_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py37h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.6.15=py37h06a4308_0
cffi=1.15.1=py37h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=37.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py37h06a4308_0
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py37h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py37h06a4308_0
jupyter=1.0.0=py37h06a4308_8
jupyter_client=7.3.5=py37h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py37h06a4308_0
jupyter_server=1.18.1=py37h06a4308_0
jupyterlab=3.4.4=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py37h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py37h295c915_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.14=0
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.5.2=py37h06a4308_0
matplotlib-base=3.5.2=py37hf590b9c_0
matplotlib-inline=0.1.6=py37h06a4308_0
mistune=0.8.4=py37h14c3975_1001
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
munkres=1.1.4=py_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py37h06a4308_0
nbconvert=6.4.4=py37h06a4308_0
nbformat=5.3.0=py37h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py37h06a4308_0
nettle=3.7.3=hbbd107a_1
notebook=6.4.12=py37h06a4308_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py37h22f2fdc_0
pip=22.1.2=py37h06a4308_0
prometheus_client=0.14.1=py37h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37_1
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py37h06a4308_0
pytorch=1.12.1=py3.7_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py37h06a4308_0
pyzmq=23.2.0=py37h6a678d5_0
qt=5.9.7=h5867ecd_1
qtconsole=5.3.2=py37h06a4308_0
qtpy=2.2.0=py37h06a4308_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=py37h06a4308_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py37h06a4308_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.2=hc218d9a_0
terminado=0.13.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
tk=8.6.11=h1ccaba5_0
torchaudio=0.12.1=py37_cu113
torchvision=0.13.1=py37_cu113
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.11=py37h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37h06a4308_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h06a4308_0
xz=5.2.5=h7f8727e_1
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py37h06a4308_0
zlib=1.2.11=0
zstd=1.4.9=haebb681_0